Clean-up of d3d12 allocator visualizer example.
